Extract sir tx vec creation to separate function
